% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_risk.R
\name{assign_risk}
\alias{assign_risk}
\title{Reducing the HES data to alcohol/tobacco-related episodes and admissions}
\usage{
assign_risk(hes, k.year.ind, youngest_age = 16, age_cat = c("16-17",
  "18-24", "25-34", "35-49", "50-64", "65-74", "75-89"),
  age_cat_start = c(16, 18, 25, 35, 50, 65, 75), method = c("Broad",
  "Narrow", "Both"), substance = c("Alcohol", "Tobacco"),
  level = c("Episode", "Admission"), summary = TRUE)
}
\arguments{
\item{hes}{cleaned HES data}

\item{k.year.ind}{year of HES data. E.g. "1718"}

\item{youngest_age}{numeric. Youngest age required in HES data}

\item{age_cat}{Character vector of age categories. Default is c("16-17", "18-24", "25-34", "35-49", "50-64", "65-74", "75-89").}

\item{age_cat_start}{Numerical vector of start ages for age categories. Default is c(16, 18, 25, 35, 50, 65, 75).}

\item{method}{method required to allocate episode to condition. "Broad" or "Narrow"}

\item{substance}{risk. "Alcohol" or "Tobacco"}

\item{level}{the level of detail of summary table. "Episode", or "Admission".}

\item{summary}{Logical - whether to return a summary table or the full HES data (want this option for the costing).
Defaults to TRUE. If TRUE, will output a summary table of HES data. If FALSE, then will output the whole year of data.}
}
\value{
Returns a summary table of number of episodes/admissions/people, SAF, and attributable episodes/admissions/people,
by age category, sex, imd_quintile, and condition. If summary = FALSE, will return the whole data unsummarised (for use in costing at episode level)
}
\description{
This is the main function used to convert the cleaned data in to summary tables of either episodes, admissions or person-specific admissions.
It reads in the cleaned Hospital episode statistics for England, reducing it to the sample selection required. It assigns an attributable fraction
to each row/episode based on either the broad or narrow method, for the substance chosen (alcohol or tobacco). The code then goes
on to calculate a summary table of alcohol/tobacco attributable hospital episodes, admissions or person-specific admissions as chosen.
}
\examples{

\dontrun{

assign_risk(hes, "1617", method = "Narrow", substance = "Alcohol", level = "Episode", summary = TRUE)

}

}
