% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignPAFs.R
\name{assignPAFs}
\alias{assignPAFs}
\title{Assign population attributable fractions to diseases}
\usage{
assignPAFs(
  data,
  paf_data = hesr::paf_data,
  paf_strat = c("ageband", "sex", "imd_quintile"),
  col_name_vec = c("diag_01")
)
}
\arguments{
\item{data}{Data table - episode-level hospital episode statistics data.}

\item{paf_data}{Data table - the population attributable fractions to 
be assigned to diseases.}

\item{paf_strat}{Character vector - the variables by which the 
population attributable fractions are stratified.}

\item{col_name_vec}{Character vector - The name of the diagnosis code columns
scanned to identify target disease diagnosis codes.}
}
\value{
Returns the input episode-level hospital episode statistics data 
with new columns added for the population attributable fractions 
that correspond to the tobacco and/or alcohol related diseases identified 
in each of the diagnosis columns.
}
\description{
Look at the data fields generated by running [hesr::findEpisodes()] 
and create a parallel set of data fields containing the 
corresponding population attributable fractions.
}
\details{
Population attributable fractions can be stratified by 
"year", sex", "ageband" and/or "imd_quintile".
}
\examples{
\dontrun{


}

}
