% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broad_method.R
\name{broad_method}
\alias{broad_method}
\title{Broad method for allocating risk \lifecycle{maturing}}
\usage{
broad_method(hes, lkup)
}
\arguments{
\item{hes}{HES data}

\item{lkup}{cleaned and formatted attributable fraction lookup table.}
}
\value{
Returns a reduced HES data, with only rows that have the largest attributable risk-related diagnosis code within an episode.
}
\description{
The broad method selects the diagnosis code of an episode to be the diagnostic code with the highest PAF.
This is in an effort to select the condition that is most causally attributable to the risk factor as to be as inclusive as possible.
}
\examples{

\dontrun{

hes <- read_hes("1718")
lkup <- format_afs("Tobacco")
broad_method(hes, lkup)

}

}
