% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cost_admission.R
\name{calc_cost_admission}
\alias{calc_cost_admission}
\title{Calculate unit costs of risk-related admissions by sex, age, IMD and condition.}
\usage{
calc_cost_admission(k.year.ind)
}
\arguments{
\item{k.year.ind}{is the year of HES data to read in.}
}
\value{
Returns an table of unit costs by condition, sex, age category and IMD quintile (where possible)
}
\description{
Calculating the cost of admissions in the latest year (2016) to assign to all years. We do this is an effort to be conservative, and cost only the episodes within
an admission that have been allocated to the primary condition. We retain only the rows in the HES data which have the same ICD10 code as the earliest risk-related
episode in the admission, and use the unit costs of episodes from the NHS Reference Costs to cost them. We sum this cost across the admission to calculate unit
admission costs.
}
\details{
We use the 'hes_cost' function to match episodes with reference costs to calculate the average cost per episode for one year of data. We found when
calculating unit costs of admissions, we are left with some subgroups that have not been costed. In order to fill in these gaps we use the function
'missing_costs' to create average unit costs of admissions by condition, sex, IMD quintile, and age category (age category depends on the size of the
data/number of admissions we have for each subgroup).
}
\examples{

\dontrun{

calc_cost_admission("1617")

}


}
