% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_hes.R
\name{clean_hes}
\alias{clean_hes}
\title{Clean HES}
\usage{
clean_hes(hes)
}
\arguments{
\item{hes}{is the raw HES data.}
}
\value{
Returns a cleaned version of the HES data.
}
\description{
This code does the basic cleaning operations on each year of HES data
}
\details{
Cleaning the data to remove any duplicates, dummy dates, missing data, and data not required (e.g. admissions that have not ended in this year)
}
\examples{

\dontrun{

hes <- read_hes("0203")
cleaned_hes <- clean_hes(hes)

}

}
