% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_hes_scot.R
\name{clean_hes_scot}
\alias{clean_hes_scot}
\title{Clean Scottish Hospital Episode Statistics}
\usage{
clean_hes_scot(hes)
}
\arguments{
\item{hes}{Data.table - is the raw Scottish HES data that is the output from \code{read_hes_scot()}..}
}
\value{
Returns a data.table that is a cleaned version of the HES data.
}
\description{
Basic cleaning operations on each year of Scottish HES data.
}
\details{
Cleaning the data to remove or consistently code 
duplicates, dummy dates, and missing data. 
Only records with data on the Index of Multiple Deprivation (IMD) decile 
in which the patient resides are retained. 
If an individual changes IMD quintile, age or sex within the year, 
then they are assigned the value for these characteristics that they had on the first admission of the year.
}
\examples{

\dontrun{

hes <- read_hes_scot("2022")
cleaned_hes <- clean_hes_scot(hes)

}

}
