% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cruk_split_oesophageal.R
\name{cruk_split_oesophageal}
\alias{cruk_split_oesophageal}
\title{CRUK Oesophageal split}
\usage{
cruk_split_oesophageal(
  hes,
  splits = hesr::cruk_splits,
  var_name = "epi_diag_selected",
  prefix = "Oesophageal"
)
}
\arguments{
\item{hes}{Data.table - the cleaned HES data.}

\item{splits}{Data.table - the distribution of C15 cases by SCC and AC}

\item{var_name}{Character string - the name of the variable containing the C15 ICD-10 code to be split}

\item{prefix}{Character string - "C15" or "Oesophageal"}
}
\value{
Returns an updated version of the HES data.table, with the diagnoses codes for 
oesophageal cancer (C15) split proportionally between Oesophageal SCC and Oesophageal AC.
}
\description{
This code splits the diagnosis codes for Oesophageal cancer (C15) 
in the HES data in to SCC or AC using CRUK splits, which are formatted and kept
in the data-raw folder `prep_cruk_splits`. 
Make sure there are no NAs in the startage field before running this. I.e. clean HES data first.
}
\examples{

\dontrun{

hes <- cruk_split_oesophageal(hes)

}

}
