% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cruk_splits.R, R/local_authority_lookups.R
\docType{data}
\name{cruk_splits}
\alias{cruk_splits}
\title{Proportion of oesophageal cases split by AC and SCC}
\format{A data table (cruk_ageband, sex, proportion_scc).}
\source{



}
\usage{
cruk_splits

cruk_splits
}
\description{
We use the CRUK splits for oesophageal cases
The data are held in the disease risk folder PR_Disease_Risk_TA/General/CRUK work/Oesophageal AC and SCC data v2.xlsx
The data is split by age groups: 1-4, 5-9, 19-14, 15-19, 20-24, 25-29, 30-34, 35-39, 40-44, 45-49, 50-54, 55-59, 60-64, 65-69, 70-74, 75-79, 80-84, 85+ and by sex.

File that contains the geographical lookup information  (from Rob) that maps ltlas to utlas
}
\keyword{datasets}
