% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_cause.R
\name{external_cause}
\alias{external_cause}
\title{External cause method \lifecycle{deprecated}}
\usage{
external_cause(hes, lkup)
}
\arguments{
\item{hes}{is the data}

\item{lkup}{is the lookup table for condition and respective ICD-10 codes to match the hes data to conditions}
}
\description{
For alcohol-related conditions, there are some that will only appear in the Cause field, rather than a diagnostic field.
These conditions will become the primary cause, over the primary diagnostic position. This would replace the function `narrow_method()`
which only looks at the primary diagnostic code.
}
\examples{

\dontrun{

external_cause(hes, hesr::lkup)

}

}
