% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findEpisodes.R
\name{findEpisodes}
\alias{findEpisodes}
\title{Assign a diagnosis to each episode}
\usage{
findEpisodes(data, lkups)
}
\arguments{
\item{data}{Data.table - the HES data.}

\item{lkups}{Data.table - the ICD-10 codes associated with tobacco and/or alcohol. 
These are stored as data within the tobalcepi package.}
}
\value{
Returns the HES data.table with a column added 
containing the ICD-10 code found during the scan.
}
\description{
Look at the first diagnostic position of all episodes within a continuous inpatient spell
 and retain episodes that have a 
relevant diagnosis in the primary diagnostic position. 
Also look for whether a relevant external cause 
has been recorded within the list of diagnoses related to that episode.
}
\details{
Diagnosis codes are identified by the function \code{icdFlag()} and 
are then synthesised to produce a single diagnosis for each episode 
(external causes where two potential diagnoses are identified). The HES 
data is then filtered to retain only episodes for which a tobacco or alcohol related 
diagnosis has been identified.
}
\examples{
\dontrun{

hes <- findEpisodes(
  data = hes,
  lkups = tobalcepi::tob_icd10_lookups
) 

}

}
