% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_afs.R
\name{format_afs}
\alias{format_afs}
\title{Clean and format attributable fractions (AFs) \lifecycle{maturing}}
\usage{
format_afs(lkup, substance = c("Alcohol", "Tobacco", "Both"))
}
\arguments{
\item{lkup}{attributable fraction lookup table}

\item{substance}{is the risk factor. E.g. Alcohol}
}
\value{
returns a table by condition, sex, icd_code, age_cat and AF))
}
\description{
Reads in lookup table of attributable fractions for either alcohol or tobacco. Formats files to be the same as each other to use in the function `assign_risk()`.
}
\details{
For now the option for 'Both' only works for the narrow method calculation. This does not include AFs, just matches the ICD-10 codes in HES to
the condition to keep the conditions for both alcohol and tobacco in order to cost them together.
}
\examples{

\dontrun{

Alcohol_AAFs <- format_afs(substance = c("Alcohol"))

}

}
