% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hes_cost.R
\name{hes_cost}
\alias{hes_cost}
\title{Cost episode-level HES data}
\usage{
hes_cost(hes, reference_costs, trimpoints, opcs_lookup, k.year)
}
\arguments{
\item{trimpoints}{Trimpoints that indicate the numbers of days in bed that are included within the standard costing. Days beyond the trimpoint need to be costed as extra.}

\item{opcs_lookup}{to add additional costs based on procedure code}

\item{k.year}{an indicator of the year that costs apply to}
}
\value{
Returns the hes data with a new column for the total episode cost (including excess trimpoints and OPCS costs), and a column for OPCS cost and trimpoints.
All inflated to 2016/17 prices.
}
\description{
Combine HES data with NHS reference costs, trimpoints, and OPCS costs.
}
\details{
Matches the HES data with the NHS reference costs, trimpoints, and OPCS costs, using the HRG code and type of admission. Reference costs,
trimpoints, and OPCS costs are stored in the package.
}
\examples{

\dontrun{

hes <- hes_cost(hes, hesr::unit_reference_costs, hesr::trimpoints, hesr::opcs_costs, k.year)

}


}
