% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_authority.R
\name{local_authority}
\alias{local_authority}
\title{Local authority lookup}
\usage{
local_authority(hes, k.year.ind)
}
\arguments{
\item{hes}{is the cleaned HES data.}
}
\value{
Returns an updated version of the HES data, adding in the up-to-date local authority identifiers
}
\description{
This code adds in the up-to-date local authority identifiers.
}
\details{
This code adds in the up-to-date local authority identifiers to fix the issue that in the data we were provided from
2002/03 to 2013/14 lower tier local authority is coded using an old version of the ONS system (that was updated in 2011). From 2014/15
The codes in the `RESLADST` field represent the old-style ONS coding system, which is no longer supported by ONS.
Old-style codes are not available for new geographies, potentially resulting in a large number now being derived as ‘Y’.
Users are advised not to use the 'RESLADST' field, but to use the `RESLADST_ONS` field instead.
`RESLADST_ONS` also represents the Local authority district of residence but uses the current ONS coding system.
}
\examples{

\dontrun{

hes_2002 <- local_authority(hes, "0203")

}

}
