% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_authority_lookups.R
\docType{data}
\name{local_authority_lookups}
\alias{local_authority_lookups}
\title{Local authority lookup table}
\format{
A data table (resladst, ltla, utla, utla_name_london_groupings, LA_name).
}
\source{

}
\usage{
local_authority_lookups
}
\description{
File that contains the geographical lookup information  (from Rob) that maps ltlas to utlas
}
\keyword{datasets}
