% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_authority_lookups.R
\docType{data}
\name{local_authority_lookups1}
\alias{local_authority_lookups1}
\title{Local authority lookup table - new ONS system}
\format{
A data table (resladst_ons, utla_name_london_groupings).
}
\source{

}
\usage{
local_authority_lookups1
}
\description{
File that contains the geographical lookup information  
that maps lower tier local authority codes to upper tier local authorities.
LTLA codes are classified by the new ONS system, e.g. "E07000004". 
For London, UTLAs are grouped into "inner" and "outer" London.
}
\keyword{datasets}
