% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_costs.R
\name{missing_costs}
\alias{missing_costs}
\title{Fill in missing costs \lifecycle{maturing}}
\usage{
missing_costs(hes, missing_cost_lookup)
}
\arguments{
\item{hes}{is the hes data after it has been ran through hes_cost.R}

\item{missing_cost_lookup}{is a lookup table to reallocate age category sizes, e.g. from 16-17 to 16-24.}
}
\value{
Returns a summary table of the unit admission costs by condition, age category, sex, and IMD quintile.
}
\description{
There will be missing unit costs where there have been zero records for a particular combination of variables.
If there are no admissions within a subgroup, we expand the age groups until there are enough admissions within a subgroup. If there are so few admissions
across all age groups, we can either: average across the condition, and allocate the same cost to everybody with that condition, or remove the condition
from the analysis. For example, for alcohol induced pseudo cushings syndrome, there was only 1 admission and therefore this is not enough to reliably
estimate a unit cost, and this condition should be removed from the analysis.
For now, We have removed the conditions that cannot be reliably costed by sex and IMD, We are currently in the process of updating this
to average across the condition (07/08/2019). This function is not finished yet.
}
\examples{

\dontrun{

hes <- hes_cost(hes, hesr::unit_reference_costs, hesr::trimpoints, hesr::opcs_costs, k.year = "2016")
hes <- missing_costs(hes, missing_cost_lookup = hesr::missing_cost_both_lookup)

}

}
