% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narrow_method.R
\name{narrow_method}
\alias{narrow_method}
\title{Narrow method for allocating risk}
\usage{
narrow_method(hes, lkup)
}
\arguments{
\item{hes}{HES data}

\item{lkup}{cleaned and formatted attributable fraction lookup table.}
}
\value{
Returns updated HES data, with only rows that have a primary risk-related diagnosis code
}
\description{
This code assigns the appropriate age-sex-condition-specific AF to each risk-related diagnosis in the *primary* diagnostic position
(non-related diagnoses are left missing) across all episodes.
}
\details{
The narrow method selects episodes that have a risk-related ICD-10 code in the primary diagnostic position or
a risk-related external cause e.g. assault. This is in an effort to select the condition that is the cause of the episode.
Primarily, the external cause is assigned the cause of the episode, and if that is missing, the primary diagnostic position is assigned. This function is used
within th function `assign_risk()`.
}
\examples{

\dontrun{

lkup <- format_afs("Tobacco")
narrow_method(hes, lkup)

}


}
