% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_data.R
\name{pop_data}
\alias{pop_data}
\title{Add the population counts by age band, sex, IMD quintile}
\usage{
pop_data(k.year, subgroup = c("sex", "imd_quintile", "age_cat"))
}
\arguments{
\item{k.year}{is the year of data}

\item{subgroup}{is the subgroup you want to collapse the data by.}
}
\value{
returns the population counts by subgroup
}
\description{
This code reads in the population data for all years, recieved from the ONS.
}
\details{
Collapses the population data by the grouping variables specified, usually sex, age group, and IMD quintile. But for the person-specific admissions, we want to
collapse by single age instead.
}
\examples{

\dontrun{

pop_data(k.year = "2016", subgroup = c("sex", "imd_quintile", "age_cat"))

}


}
