% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_selection.R
\name{sample_selection}
\alias{sample_selection}
\title{Reduce size of HES to sample required}
\usage{
sample_selection(hes, start_age = 16, age_categories = c("16-17",
  "18-24", "25-34", "35-49", "50-64", "65-74", "75-89"),
  age_cat_start_age = c(16, 18, 25, 35, 50, 65, 75))
}
\arguments{
\item{hes}{cleaned HES data}
}
\value{
Returns reduced HES data,
}
\description{
Reads in cleaned Hospital episode statistics and reduces to include only sample we need, e.g. reduced ages, and assigns admission types.
}
\details{
This code reduces the size of the HES data to only include applicable rows. This function is used within the function `assign_risk()`.
}
\examples{

\dontrun{

sample_selection(hes,
                 start_age = 16,
                 age_categories = c("16-17", "18-24", "25-34", "35-49", "50-64", "65-74", "75-89"),
                 age_cat_start_age = c(16, 18, 25, 35, 50, 65, 75)
                 )

}


}
