% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_year_reference_costs.R
\name{single_year_reference_costs}
\alias{single_year_reference_costs}
\title{Format and clean Reference Costs for single year}
\usage{
single_year_reference_costs(k.year.ind)
}
\arguments{
\item{k.year.ind}{is the year of HES data to read in.}
}
\value{
Returns a table of episode unit costs by ICD10 code and admission type.
}
\description{
Combining the reference costs for different types of admission. We only run this if we want one year of reference costs, such as 2016.
We do not use this because we find it leaves many HRG codes unmatched. But this function can be used as a test. Instead we prepare the unit reference costs for all years in the folder `data-raw` in
 `unit_reference_costs`. They are stored in the package in hesr::unit_reference_costs.
}
\examples{

\dontrun{

single_year_reference_costs("1617")

}


}
