% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimpoints.R
\docType{data}
\name{trimpoints}
\alias{trimpoints}
\title{Format and clean Trimpoints}
\format{
A data table (sushrg, Episode.Trimpoint).
}
\source{

}
\usage{
trimpoints
}
\description{
We use the Trimpoints provided by NHS Digital
Trimpoints are used to define a threshold: bed days occurring within the trimpoint (truncated bed days), and bed days occurring above the trimpoint (excess bed days)
are reported, costed and priced separately.
}
\details{
Trimpoints are calculated from episodes in the HES APC extract for a previoud financial year (the trimpoints used for 2017/18 are those derived from length of stay
in 2016/17 activity)
}
\keyword{datasets}
