% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_reference_costs.R
\docType{data}
\name{unit_reference_costs}
\alias{unit_reference_costs}
\title{Format and clean reference costs}
\format{
A data table (sushrg, description, cost, Excess_cost, Year, type).
}
\source{

}
\usage{
unit_reference_costs
}
\description{
We use the NHS Digital Reference costs for years 2010-2016 to include as many HRG codes as possible, as when we used 2016 reference costs, we found ~70% remained unmatched.
Combining the reference costs for different types of admission.
Returns an table of episode unit costs by ICD10 code and admission type.
}
\keyword{datasets}
