% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cruk_split_oesophageal.R
\name{cruk_split_oesophageal}
\alias{cruk_split_oesophageal}
\title{Split Oesophageal cancer cases into AC and SCC subtypes}
\usage{
cruk_split_oesophageal(
  data,
  splits = hesr::cruk_splits,
  col_name_vec = c("diag_01"),
  prefix = "Oesophageal"
)
}
\arguments{
\item{data}{Data table - the cleaned hospital episode statistics data.}

\item{splits}{Data table - the distribution of C15 cases by SCC and AC}

\item{col_name_vec}{Character vector - The name of the diagnosis code columns
scanned to identify target disease diagnosis codes.}

\item{prefix}{Character string - "C15" or "Oesophageal"}
}
\value{
Returns an updated version of the input hospital episode statistics data, 
with the diagnoses codes for 
oesophageal cancer (C15) split proportionally between Oesophageal SCC and Oesophageal AC.
}
\description{
This code splits the diagnosis codes for Oesophageal cancer (C15) 
in the hospital episode statistics data in to SCC or AC.
}
\details{
The percentage splits are informed by estimates given to us by Cancer Research UK, 
which are formatted and embedded in the package as [hesr::cruk_splits].

Make sure there are no missing values in the startage field before running this.

Run on the episode-level data after running [hesr::findEpisodes()].
}
\examples{

\dontrun{


}

}
