% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cruk_split_oesophageal.R
\name{cruk_split_oesophageal}
\alias{cruk_split_oesophageal}
\title{CRUK Oesophageal split \lifecycle{maturing}}
\usage{
cruk_split_oesophageal(hes)
}
\arguments{
\item{hes}{is the cleaned HES data.}
}
\value{
Returns an updated version of the HES data, with the diagnoses codes for oesophageal cancer (C15) split proportionally between Oesophageal SCC and Oesophageal AC.
}
\description{
This code splits the diagnosis codes for Oesophageal cancer (C15) in the HES data in to SCC or AC using CRUK splits, which are formatted and kept
in the data-raw folder `prep_cruk_splits`. Make sure there are no NAs in the startage field before running this. I.e. clean HES data first. This function is used
within the function `narrow_method()`.
}
\examples{

\dontrun{

hes <- cruk_split_oesophageal(hes)

}

}
