% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findEpisodes.R
\name{findEpisodes}
\alias{findEpisodes}
\title{Search for episodes with the target International Classification of Disease diagnoses}
\usage{
findEpisodes(data, lkups, col_name_vec, external_cause_names)
}
\arguments{
\item{data}{Data table - episode-level hospital episode statistics data.}

\item{lkups}{Data table - the International Classification of Disease 
codes associated with tobacco and/or alcohol related conditions.}

\item{col_name_vec}{Character vector - the column names to be scanned for matching disease diagnosis codes.}

\item{external_cause_names}{Character vector - the names of partially attributable acute conditions 
that will be used to identify whether the episode was assigned an external cause.}
}
\value{
Returns the input episode-level hospital episode statistics data 
with a new column added for each value of col_name_vec
containing the name of the condition found during the scan.
}
\description{
Look at specified diagnostic positions of episodes
and create a new column in the dataset that picks out the selected episodes.
}
\details{
Diagnosis codes are identified by the function [hesr::icdFlag()], 
which is called by this function. This function loops through the values of 
col_names_vec applying [hesr::icdFlag()].

Uses the function [tobalcepi::ExpandCodes()] to convert the disease code lookup tables 
within the tobalcepi package into long form, ready for use in the matching process.
}
\examples{
\dontrun{


}

}
