% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hesr-package.R
\docType{package}
\name{hesr-package}
\alias{hesr}
\alias{hesr-package}
\title{hesr: Process Hospital Episode Statistics Data}
\description{
A collection of functions to read, clean and analyse admitted patient care hospital episode statistics data to inform the health economic modelling of tobacco and alcohol policy.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://stapm.gitlab.io/r-packages/hesr/}
}

}
\author{
\strong{Maintainer}: Duncan Gillespie \email{duncan.gillespie@sheffield.ac.uk} (\href{https://orcid.org/0000-0003-3450-5747}{ORCID})

Authors:
\itemize{
  \item Laura Webster
  \item Colin Angus \email{c.r.angus@sheffield.ac.uk}
  \item Alan Brennan \email{a.brennan@sheffield.ac.uk}
}

}
\keyword{internal}
