% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_authority_lookups.R
\docType{data}
\name{local_authority_lookups}
\alias{local_authority_lookups}
\title{Local authority lookup table - old ONS system}
\format{
A data table (resladst, utla_name_london_groupings).
}
\usage{
local_authority_lookups
}
\description{
File that contains the geographical lookup information  
that maps lower tier local authority codes to upper tier local authorities.
LTLA codes are classified by the old system, e.g. "00AA". 
For London, UTLAs are grouped into "inner" and "outer" London.
}
\keyword{datasets}
