% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hes_scot.R
\name{read_hes_scot}
\alias{read_hes_scot}
\title{Read the Scottish Hospital Episode Statistics data files}
\usage{
read_hes_scot(k.year.ind)
}
\arguments{
\item{k.year.ind}{Integer - the year of Scottish HES data. In the format "2022".}
}
\value{
Returns a data table. Note that:
\itemize{
\item All variable names are converted to lower case.
}
}
\description{
Reads in Hospital Episode Statistics for Scotland data files for each year from 2007/08 to 2021/22. 
The data were provided with only the variables and disease codes needed for computing 
the rates of diseases related to tobacco and/or alcohol and the associated rates and costs 
of hospital admissions.
}
\details{
This code reads in the data for one year and produces a cleaned data file ready for further analysis.

Variables included

\itemize{
  \item PATIENT_ID
  \item AGE_IN_YEARS
  \item SEX - 1 = male, 2 = female. 0 or 9 are not known and not specified.
  \item ADMIDATE
  \item ADMISSION_TYPE - `https://www.ndc.scot.nhs.uk/Dictionary-A-Z/Definitions/index.asp?Search=A&ID=58&Title=Admission%20Type`
  \item ADMISSION_TRANSFER_FROM- `https://www.ndc.scot.nhs.uk/Dictionary-A-Z/Definitions/index.asp?Search=A&ID=59&Title=Admission/Transfer%20From`
  \item DISDATE
  \item DISCHARGE_TRANSFER_TO - `https://www.ndc.scot.nhs.uk/Dictionary-A-Z/Definitions/index.asp?Search=D&ID=228&Title=Discharge/Transfer%20To%20-%20Location`
  \item DISCHARGE_TYPE - `https://www.ndc.scot.nhs.uk/Dictionary-A-Z/Definitions/index.asp?Search=D&ID=224&Title=Discharge%20Type`
  \item SPELBGIN - 0 = Not first episode of spell; 1 = First episode of spell that started in previous year; 2 = First episode of spell that started in current year
  \item DISCHARGE_DATE
  \item ADMISSION_DATE
  \item SPELDUR - the difference in days between the admission date (admidate) and the discharge date (epiend) provided the discharge method (dismeth) confirms that the spell has finished. If the episode has not finished it is calculated from the end of the year and admidate.
  \item SPELEND - whether the episode is the last of a spell. It is set for finished episodes (episode status - epistat - is 3) for general, delivery or birth episodes (episode type - epitype - is 1, 2 or 3) provided the discharge method (dismeth) confirms that the spell has finished. Y = Last episode of spell ; N = Not last episode of spell
  \item LENGTH_OF_STAY - The patients length of stay derived from admission date and discharge date.
  \item EPISODE_MARKER - indicating the order of episodes of care
  \item GLS_CIS_MARKER - SMR01 Continuous Inpatient Stay Marker (including GLS records)
  \item MAIN_CONDITION - `https://www.ndc.scot.nhs.uk/Data-Dictionary/SMR-Datasets/General-Clinical-Information/Diagnostic-Section/Main-Condition.asp`
  \item OTHER_CONDITION_1 - `https://www.ndc.scot.nhs.uk/Dictionary-A-Z/Definitions/index.asp?Search=O&ID=365&Title=Other%20Condition/Co-morbidity%20and%20Complication%20ICD10%20(2%20-%206)`
  \item OTHER_CONDITION_2
  \item OTHER_CONDITION_3
  \item OTHER_CONDITION_4
  \item OTHER_CONDITION_5
  \item ADMISSION_REASON - `https://www.ndc.scot.nhs.uk/Data-Dictionary/SMR-Datasets/Episode-Management/Admission-Reason/`
  \item MAIN_OPERATION - `https://www.ndc.scot.nhs.uk/Dictionary-A-Z/Definitions/index.asp?Search=M&ID=316&Title=Main%20Operation/Treatment/Investigative%20Procedure/Intervention%20-%20OPCS4`
  \item OTHER_OPERATION_1 - `https://www.ndc.scot.nhs.uk/Dictionary-A-Z/Definitions/index.asp?Search=O&ID=368&Title=Other%20Operation/Treatment/Intervention/Investigative%20Procedures%20(2-4)%20-%20OPCS4`
  \item OTHER_OPERATION_2
  \item OTHER_OPERATION_3
  \item MANAGEMENT_OF_PATIENT - `https://www.ndc.scot.nhs.uk/Dictionary-A-Z/Definitions/index.asp?Search=O&ID=368&Title=Other%20Operation/Treatment/Intervention/Investigative%20Procedures%20(2-4)%20-%20OPCS4`
  \item SPECIALTY - `https://www.ndc.scot.nhs.uk/Dictionary-A-Z/Definitions/index.asp?Search=S&ID=473&Title=Specialty/Discipline`
  \item HRG42 - Healthcare Resource Group (HRG) codes are standard groupings of clinically similar treatments which use comparable levels of healthcare resource. The field contains version 4 HRG codes. 
  \item PROVIDER_CODE - `https://www.ndc.scot.nhs.uk/Dictionary-A-Z/Definitions/index.asp?Search=P&ID=932&Title=Provider%20Code`
  \item simd2016_sc_quintile - Scottish Index of Multiple Deprivation quintile of patient residence (2016 version). Least deprived = 5, most deprived = 1
  \item EPISODE_RECORD_KEY - sequential number generated by PAS identifying the episode
}
}
\examples{

\dontrun{

hes_2002 <- read_hes("0203", test = TRUE)

}

}
