% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_selection.R
\name{sample_selection}
\alias{sample_selection}
\title{Reduce size of HES to sample required}
\usage{
sample_selection(hes, start_age, age_categories, age_cat_start_age)
}
\arguments{
\item{hes}{data.table - the cleaned HES data.}

\item{start_age}{Integer - the minimum age to consider for non-maternity admissions.}

\item{age_categories}{Character vector - the age category labels.}

\item{age_cat_start_age}{Integer vector - the ages that open each age category.}
}
\value{
Returns a data.table of reduced HES data.
}
\description{
Reads in cleaned Hospital episode statistics and reduces to include only sample we need, 
e.g. reduced ages, and assigns admission types.
}
\details{
This code reduces the size of the HES data to only include applicable rows. 
See the \href{https://stapm.gitlab.io/r-packages/hesr/articles/cleaning_and_costing.html#sample-selection}{cleaning and costing vignette} for details.
}
\examples{

\dontrun{

sample_selection(hes,
                 start_age = 16,
                 age_categories = c("16-17", "18-24", "25-34", "35-49", "50-64", "65-74", "75-89"),
                 age_cat_start_age = c(16, 18, 25, 35, 50, 65, 75)
                 )

}


}
